+++ packages/gql/src/balancer-gauges/index.server.ts
@@ -6,7 +6,7 @@ import { getSdk as gnosisSdk } from "./__generated__/Gnosis.server";
 import { getSdk as goerliSdk } from "./__generated__/Goerli.server";
 import { getSdk as optimismSdk } from "./__generated__/Optimism.server";
 import { getSdk as polygonSdk } from "./__generated__/Polygon.server";
-import { getSdk as polygonZkEVMSdk } from "./__generated__/Polygonzkevm.server";
+import { getSdk as polygonZkEVMSdk } from "./__generated__/Polygon-zkevm.server";
 import { getSdk as sepoliaSdk } from "./__generated__/Sepolia.server";
 
 export default {
