+++ packages/gql/src/balancer-pools/index.ts
@@ -6,7 +6,7 @@ import { getSdkWithHooks as gnosisSdk } from "./__generated__/Gnosis";
 import { getSdkWithHooks as goerliSdk } from "./__generated__/Goerli";
 import { getSdkWithHooks as optimismSdk } from "./__generated__/Optimism";
 import { getSdkWithHooks as polygonSdk } from "./__generated__/Polygon";
-import { getSdkWithHooks as polygonZkEVMSdk } from "./__generated__/Polygonzkevm";
+import { getSdkWithHooks as polygonZkEVMSdk } from "./__generated__/Polygon-zkevm";
 import { getSdkWithHooks as sepoliaSdk } from "./__generated__/Sepolia";
 
 export default {
